move _winnow_process_list to utils/multiprocessing_utils.py
