+++ src/ophys_etl/modules/segmentation/merge/characteristic_timeseries.py
@@ -13,7 +13,7 @@ else:
 import numpy as np
 import multiprocessing
 import multiprocessing.managers
-from ophys_etl.modules.segmentation.merge.utils import (
+from ophys_etl.modules.segmentation.utils.multiprocessing_utils import (
     _winnow_process_list)
 
 from ophys_etl.modules.segmentation.\
